+++ cmd/cmd_test.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"errors"
 	"os"
 	"strings"
 	"testing"
@@ -11,6 +12,15 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+var now = time.Now()
+var FullLog = []entry.Entry{
+	{Time: now, Category: "First", Note: "First"},
+	{Time: now, Category: "Second", Note: "Second"},
+	{Time: now, Category: "Third", Note: "Third"},
+}
+
+var EmptyLog = []entry.Entry{}
 
 func AssertEqualSlice(t *testing.T, expected, actual []entry.Entry) {
 	t.Helper()
@@ -104,5 +114,130 @@ func newFile(t *testing.T, entries []entry.Entry) string {
 	return writer.Name()
 }
 
-func TestCommand(t *testing.T) {
+type Expected struct {
+	Entry  entry.Entry
+	Err    error
+	Stdout string
+	Stderr string
+}
+
+type Test struct {
+	Name string
+	Args []string
+	// GOTIME_LOG is set **always** to contain LogState
+	Envs     map[string]string
+	LogState []entry.Entry
+	Expected Expected
+}
+
+func testCmd(t *testing.T, test Test) {
+	t.Run(test.Name, func(t *testing.T) {
+		tc := NewTestConfig(t).
+			WithEnvs(test.Envs).
+			WithLogFile(test.LogState)
+		config := tc.Config()
+
+		err := Run(test.Args, config)
+
+		require := require.New(t)
+		if errors.Is(test.Expected.Err, assert.AnError) {
+			require.Error(err)
+			return
+		} else if test.Expected.Err != nil {
+			require.ErrorIs(err, test.Expected.Err)
+			return
+		}
+
+		require.Nil(err)
+		assert.Equal(t, test.Expected.Stdout, tc.Stdout())
+		assert.Equal(t, test.Expected.Stderr, tc.Stderr())
+
+		entries, err := config.GetAll()
+		require.Nil(err)
+		if len(entries) != 0 {
+			entries = entries[:len(entries)-1]
+		}
+		AssertEqualSlice(t, test.LogState, entries)
+	})
+}
+
+func TestCmds(t *testing.T) {
+	tests := []Test{
+		{
+			Name: "default/valid input/empty file",
+			Args: []string{"happy", "Cool", "note"},
+			Expected: Expected{
+				Entry: entry.Entry{Category: "happy", Note: "Cool note"},
+			},
+		},
+		{
+			Name: "default/valid input/existing file",
+			Args: []string{"second", "Note"},
+			LogState: []entry.Entry{
+				{Time: now, Category: "first", Note: "Should remain unchanged"},
+			},
+			Expected: Expected{
+				Entry: entry.Entry{Category: "second", Note: "Note"},
+			},
+		},
+		{
+			Name: "log/valid input/empty file",
+			Args: []string{"log", "happy", "Cool", "note"},
+			Expected: Expected{
+				Entry: entry.Entry{Category: "happy", Note: "Cool note"},
+			},
+		},
+		{
+			Name: "log/valid input/existing file",
+			Args: []string{"log", "second", "Note"},
+			LogState: []entry.Entry{
+				{Time: now, Category: "first", Note: "Should remain unchanged"},
+			},
+			Expected: Expected{
+				Entry: entry.Entry{Category: "second", Note: "Note"},
+			},
+		},
+		{
+			Name: "append/valid input/valid file",
+			Args: []string{"append", "Cool", "note"},
+			LogState: []entry.Entry{
+				{Time: now, Category: "current", Note: "Should remain unchanged"},
+			},
+			Expected: Expected{
+				Entry: entry.Entry{Category: "current", Note: "Cool note"},
+			},
+		},
+		{
+			Name: "append/empty file",
+			Args: []string{"append", "Cool", "note"},
+			Expected: Expected{
+				Err: ErrFileEmpty,
+			},
+		},
+		{
+			Name: "help/short",
+			Args: []string{"-h"},
+			Expected: Expected{
+				Err: assert.AnError,
+			},
+		},
+		{
+			Name: "help/long",
+			Args: []string{"--help"},
+			Expected: Expected{
+				Err: assert.AnError,
+			},
+		},
+		{
+			Name: "bad flag",
+			Args: []string{"--badflag123"},
+			Expected: Expected{
+				Err: assert.AnError,
+			},
+		},
+	}
+
+	for _, test := range tests {
+		testCmd(t, test)
+	}
 }